:truck: move `TierIconBox` component from `imagebox` to `tier` #32
