+++ src/components/frame/with-buttons/menu-button/game-info/TierInfo.tsx
@@ -3,7 +3,7 @@ import { indexProps } from './types';
 
 import SettingNavigationButton from '@/components/button/SettingNavigationButton';
 import { HIGH_TIER_INFO, LOW_TIER_INFO } from '@/components/imagebox/constants';
-import TierBox from '@/components/imagebox/TierIconBox';
+import TierBox from '@/components/tier';
 
 const TierInfo = ({ onBack, onClose }: indexProps) => {
   return (
