+++ src/pages/play/lobby/Players/Info.tsx
@@ -2,7 +2,7 @@ import * as styles from './index.css';
 
 import type { Player } from '@/types/play';
 
-import TierIconBox from '@/components/imagebox/TierIconBox';
+import TierIconBox from '@/components/tier';
 
 interface Props {
   player: Player;
